# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MUROTECH SaaS API
  description: Sistema de Facturación Electrónica para Costa Rica (API MH v4.4)
  version: 1.0.0
  contact:
    name: MUROTECH
    url: https://murotech.com
servers:
  - url: http://localhost:5001
    description: Desarrollo local
  - url: https://murotechsaas-95ru.onrender.com
    description: Producción
...
paths:
...
  /api/pagos/checkout:
    post:
      tags: [Pagos]
      summary: Crear orden de pago
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [empresa_id]
              properties:
                empresa_id: { type: string }
                plan_tipo: { type: string, enum: [basico, emisor, premium, enterprise] }
      responses:
        '200': { description: Orden de pago creada }
        '404': { description: Empresa no encontrada }
  /api/pagos/confirmar:
    post:
      tags: [Pagos]
      summary: Confirmar pago
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_id]
              properties:
                payment_id: { type: string }
                provider: { type: string }
                transaction_id: { type: string }
      responses:
        '200': { description: Pago confirmado }
        '400': { description: Pago ya procesado }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...